requirements.txt
torch>=2.1
transformers>=4.37
Pillow>=10.2
redis>=5.0
requests>=2.31
python-dotenv>=1.0
insightface>=0.7
opencv-python-headless>=4.9
numpy>=1.26
fastapi>=0.115
uvicorn>=0.30
timm>=0.9
